requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
certifi=2018.11.29=py36_0
click=7.0=pypi_0
django=2.1.5=pypi_0
django-disqus=0.5=pypi_0
django-tagging=0.4.6=pypi_0
django-widget-tweaks=1.4.3=pypi_0
flask=1.0.2=pypi_0
itsdangerous=1.1.0=pypi_0
jinja2=2.10=pypi_0
markupsafe=1.1.0=pypi_0
mock=2.0.0=pypi_0
pbr=5.1.1=pypi_0
pillow=5.4.1=pypi_0
pip=18.1=py36_0
python=3.6.8=h9f7ef89_0
pytz=2018.9=pypi_0
setuptools=40.6.3=py36_0
six=1.12.0=pypi_0
sqlite=3.26.0=he774522_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.15.26706=h3a45250_0
werkzeug=0.14.1=pypi_0
wheel=0.32.3=py36_0
wincertstore=0.2=py36h7fe50ca_0
